+++ js/modules/fullSize-picture.js
@@ -1,4 +1,5 @@
 import {isEscapeKey} from './utils.js';
+import { renderComments, clearComments } from './render-comments.js';
 
 const bigPictureModalElement = document.querySelector('.big-picture');
 const closeBigPictureModalElement = bigPictureModalElement.querySelector('.big-picture__cancel');
@@ -8,32 +9,11 @@ const bigPictureShownComments = bigPictureModalElement.querySelector('.social__c
 const bigPictureTotalComments = bigPictureModalElement.querySelector('.social__comment-total-count');
 const bigPictureDescription = bigPictureModalElement.querySelector('.social__caption');
 
-
-const template = document.querySelector('#comment').content.querySelector('.social__comment');
-const commentListElement = document.querySelector('.social__comments');
-
-const socialCommentCount = bigPictureModalElement.querySelector('.social__comment-count');
-const commentLoader = bigPictureModalElement.querySelector('.comments-loader');
-
 const onDocumentKeydown = (evt) => {
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     closeBigPictureModal();
   }
-};
-
-const renderComments = (comments) => {
-  comments.forEach((comment) => {
-    const templateComment = template.cloneNode(true);
-    const image = templateComment.querySelector('.social__picture');
-
-    image.src = comment.avatar;
-    image.alt = comment.name;
-
-    templateComment.querySelector('.social__text').textContent = comment.message;
-
-    commentListElement.appendChild(templateComment);
-  });
 };
 
 const renderBigPictureModalElement = (photo) => {
@@ -43,6 +23,7 @@ const renderBigPictureModalElement = (photo) => {
   bigPictureTotalComments.textContent = photo.comments.length;
   bigPictureDescription.textContent = photo.description;
 
+  clearComments();
   renderComments(photo.comments);
 };
 
@@ -66,13 +47,5 @@ function closeBigPictureModal() {
 closeBigPictureModalElement.addEventListener('click', () => {
   closeBigPictureModal();
 });
-
-// Hide commens & comments count
-const socialCommentDelete = () => {
-  socialCommentCount.classList.add('hidden');
-  commentLoader.classList.add('hidden');
-};
-
-socialCommentDelete();
 
 export { openBigPictureModal };
